move fetcher to backend
