--- fetcher/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "youtoop-guide",
-  "version": "1.0.0",
-  "main": "index.js",
-  "repository": "git@github.com:wiennat/youtoop-guide.git",
-  "author": "wiennat <wiennat@gmail.com>",
-  "license": "MIT",
-  "scripts": {
-    "fetch:data": "node -r dotenv/config fetch.js data",
-    "fetch:filter": "node -r dotenv/config fetch.js filter"
-  },
-  "dependencies": {
-    "chalk": "^2.4.2",
-    "dotenv": "^6.2.0",
-    "isomorphic-fetch": "^2.2.1",
-    "knex": "^0.16.3",
-    "mysql": "^2.16.0"
-  }
-}
